refactor: separate files to folders
Also fix login button font
